pyproject.toml
[tool.poetry]
name = "emkademy"
version = "0.1.0"
description = ""
authors = ["CanerOren <canerorenn@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.15"
pandas = "~=2.3"
matplotlib = "~=3.10"
pydantic = "^2"
hydra-core = "^1.3"
fsspec = {version = "~=2025.10", extras = ["gcs"]}
gcsfs = "~=2025.10"
torch = "==2.9.1"
lightning = "==2.5.6"
torchmetrics = ">=1.0.0,<2.0.0"
transformers = {version= "~=4.57", extras = ["torch"]}
mlflow = "==3.6.0"
psycopg2-binary = "~=2.9"
google-cloud-compute = "~=1.40"

[tool.poetry.group.dev.dependencies]
jupyterlab = "~=4.5"
pytest = "~7.3"
black= "*"
isort= "*"
flake8= "*"
mypy= "*"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length=120

[tool.isort]
profile="black"
multi_line_output= 3 
include_trailing_comma=true
force_grid_wrap= 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 120
indent = 4
lines_between_types = 1
known_first_party= ["cybulde"]
default_section = "THIRDPARTY"
sections= ["FUTURE","STDLIB","THIRDPARTY","FIRSTPARTY","LOCALFOLDER"]

[tool.mypy]
python_version = "3.12"
show_error_codes = true
ignore_missing_imports = true
install_types= true
non_interactive = true
follow_imports = "silent"
no_strict_optional = true

warn_redundant_casts = false
warn_unused_ignores = true
warn_unused_configs = true
warn_return_any = true
warn_no_return = true
warn_incomplete_stub= true

disallow_subclassing_any = true

disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true
check_untyped_defs = true
strict_optional = true

verbosity = 0
